Use explicit `uint256` in WETH9 event declarations

+++ WETH9.sol
@@ -20,10 +20,10 @@ contract WETH9 {
     string public symbol   = "WETH";
     uint8  public decimals = 18;
 
-    event  Approval(address indexed src, address indexed guy, uint wad);
-    event  Transfer(address indexed src, address indexed dst, uint wad);
-    event  Deposit(address indexed dst, uint wad);
-    event  Withdrawal(address indexed src, uint wad);
+    event  Approval(address indexed src, address indexed guy, uint256 wad);
+    event  Transfer(address indexed src, address indexed dst, uint256 wad);
+    event  Deposit(address indexed dst, uint256 wad);
+    event  Withdrawal(address indexed src, uint256 wad);
 
     mapping (address => uint)                       public  balanceOf;
     mapping (address => mapping (address => uint))  public  allowance;
